Add stage_senders header
