+++ node/silkworm/stagedsync/stage_senders.cpp
@@ -14,11 +14,7 @@
    limitations under the License.
 */
 
-#include <silkworm/common/log.hpp>
-#include <silkworm/db/stages.hpp>
-#include <silkworm/stagedsync/recovery/recovery_farm.hpp>
-
-#include "stagedsync.hpp"
+#include "stage_senders.hpp"
 
 namespace silkworm::stagedsync {
