+++ node/silkworm/stagedsync/stagedsync.hpp
@@ -32,20 +32,6 @@
 
 namespace silkworm::stagedsync {
 
-class Senders final : public IStage {
-  public:
-    explicit Senders(NodeSettings* node_settings) : IStage(db::stages::kSendersKey, node_settings){};
-    ~Senders() override = default;
-
-    StageResult forward(db::RWTxn& txn) final;
-    StageResult unwind(db::RWTxn& txn, BlockNum to) final;
-    StageResult prune(db::RWTxn& txn) final;
-    std::vector<std::string> get_log_progress() final;
-    bool stop() final;
-
-  private:
-    std::unique_ptr<recovery::RecoveryFarm> farm_{nullptr};
-};
 
 class Execution final : public IStage {
   public:
