+++ node/silkworm/stagedsync/sync_loop.cpp
@@ -19,6 +19,7 @@
 #include <boost/format.hpp>
 
 #include <silkworm/stagedsync/stage_blockhashes.hpp>
+#include <silkworm/stagedsync/stage_senders.hpp>
 #include <silkworm/stagedsync/stagedsync.hpp>
 
 namespace silkworm::stagedsync {
